examples/minio/resize/requirements.txt
kubernetes
minio
Pillow
